pyproject.toml
[project]
name = "kpe-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.3.2",
    "litellm>=1.63.3",
    "numpy>=2.2.3",
    "openai>=1.65.4",
    "openrouter>=1.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "sentence-transformers>=3.4.1",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
]
